Split out base view classes into their own modules
