+++ timetable_solutions/tests/unit/interfaces/data_management/views/test_search_view_base.py
@@ -1,10 +1,10 @@
-from interfaces.data_management.views import base_views
+from interfaces.data_management.views import _base_search_view
 
 
 class TestSearchViewProperties:
     """Tests for the properties defined on the SearchView class."""
 
-    class FakeSearchView(base_views.SearchView):
+    class FakeSearchView(_base_search_view.SearchView):
         displayed_fields = {
             "test_field_one": "Test Field One",
             "test_field_two": "Test Field Two",
